+++ src/ext/vers.rs
@@ -1,8 +1,8 @@
 use crate::binary::RankerT;
 
 impl RankerT for vers_vecs::RsVec {
-    fn new_packed(seq: &[usize]) -> Self {
-        let bv = vers_vecs::BitVec::from_limbs_iter(seq.iter().map(|x| !*x as u64));
+    fn new_packed(seq: &[u64]) -> Self {
+        let bv = vers_vecs::BitVec::from_limbs_iter(seq.iter().map(|x| !*x));
         Self::from_bit_vec(bv)
     }
